requirements.txt
pyTelegramBotAPI
google-generativeai
aiohttp
md2tgmd
unidecode
python-dotenv